Added Shader class
Added shader class, which needs to be provided with vertex shader source and fragment shader source (as string).
In the demo scene, shaders are set up to render a red triangle.

+++ Game/Game/Engine/source/MyMath.cpp
@@ -2,12 +2,12 @@
 
 
 
-Vector2::Vector2(double x, double y){
+Vector2::Vector2(float x, float y){
 	this->x = x;
 	this->y = y;
 }
 
-Vector3::Vector3(double x, double y, double z){
+Vector3::Vector3(float x, float y, float z){
 	this->x = x;
 	this->y = y;
 	this->z = z;
